backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kami-sim"
version = "0.1.0"
description = "LLM-driven multi-agent town simulation - backend"
requires-python = ">=3.11"
dependencies = [
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "fastapi>=0.115",
    "uvicorn>=0.32",
    "anthropic>=0.40",
    "networkx>=3.2",
    "pydantic>=2.9",
    "chromadb>=0.5",
    "websockets>=13.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
